solucione bugs refactorice e hice los css de los componentes del dashboard
