+++ client/src/Components/DashBoard/BarberCard/BarberCard.css
@@ -1,0 +1,32 @@
+.barberCardDetail{
+    display:flex;
+    flex-direction: column;
+    height: 550px;
+    width: 350px;
+    background-color: aquamarine;
+    border: 5px solid rgba(0, 0, 0, 0.273);
+    border-radius: 8px;
+    padding: 15px;
+    align-items: center;
+    justify-content: center;
+    box-sizing: border-box;
+
+}
+
+.barberCardDetail img{
+    position: relative;
+    width: 50%;
+    margin-bottom: 10%;
+}
+
+.barberCardDetail2{
+    border: 1px solid rgba(0, 0, 0, 0.36);
+    border-radius: 8px;
+    width: 85%;
+    padding: 10px;
+    box-sizing: border-box;
+}
+
+.barberCardDetail2 button {
+    margin: 5px;
+}
